// src/common/models/modal-view.ts
export enum IModalView {
  redeemInfo = 'redeemInfo',
  redeemConfirm = 'redeemConfirm',
  redeemConfirmed = 'redeemConfirmed',
  withdrawInfo = 'withdrawInfo',
  withdrawInstallWallet = 'withdrawInstallWallet',
  withdrawConnectWallet = 'withdrawConnectWallet',
  withdrawSuccess = 'withdrawSuccess',
  withdrawConfirm = 'withdrawConfirm',
}